requirements.txt
fastapi
uvicorn
tensorflow
pillow
opencv-python
scikit-learn
simpleitk
matplotlib
numpy
pandas
